requirements.txt
pysondb==1.6.7
google-cloud-texttospeech==2.11.0
google-cloud-storage==2.3.0
Flask==2.1.2
flask-restful==0.3.9
flask-cors==3.0.10
gunicorn==20.1.0
